refactor: 사용자 이름 입력 화면에서 사용하는 코드 및 리소스를 :feature:username으로 옮김
