+++ feature/username/src/main/java/com/practice/username/RegisterFormUiState.kt
@@ -1,0 +1,16 @@
+package com.practice.username
+
+data class RegisterFormUiState(
+    val name: String,
+    val isDuplicateName: Boolean,
+) {
+    val isNameValid: Boolean
+        get() = NameValidator.validate(name)
+
+    companion object {
+        val Empty = RegisterFormUiState(
+            name = "",
+            isDuplicateName = false,
+        )
+    }
+}
