+++ feature/username/src/main/res/values/strings.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="register_form_screen">닉네임 입력</string>
+    <string name="next_button">다음</string>
+
+    <string name="name_title">이름</string>
+    <string name="name_placeholder">한글 15자, 영문, 숫자는 30자까지 가능합니다.</string>
+    <string name="username_invalid">조건에 맞지 않는 이름입니다.</string>
+    <string name="username_duplicate">중복된 사용자 이름입니다.</string>
+</resources>
